source-code/worker/requirements.txt
sqlalchemy
psycopg2-binary
pika
boto3
pillow
python-dotenv
requests